nova ceph: just read the fsid out of node attributes
We were trying to grep it from ceph output, which was making
all kinds of mess in terms of getting the data early enough. But
as long as we're using that static info, we might as well grab it
from the original source -- the node["ceph"] attributes which we
already know are present.